% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casualties_lookup}
\alias{casualties_lookup}
\title{Stats19 schema and variables}
\description{
\code{casualties_lookup} contains a lookup table for recoding casualty types.
}
\keyword{datasets}
