% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{cluster_junction}
\alias{cluster_junction}
\title{Cluster junction points into polygons}
\usage{
cluster_junction(x, dist = 15, nQuadSegs = 3)
}
\arguments{
\item{x}{a SF data frame of joints}

\item{dist}{buffer distance past to sf::st_buffer}

\item{nQuadSegs}{how many segments per quatrant? 5 is the default.}
}
\value{
Returns an SF data frame of POLYGONS
}
\description{
Cluster junction points into polygons
}
\details{
This function clusters points together and defines junction
polygons, the size of the polygons is dictated by \code{dist}. For single
junction points a circle around the junction point is returned. For
clustered junctions a polygon enclosing the whole junction area is
returned. A column called junction_ids provides a looup list between the
junction clusters and the junction points.
}
\examples{
x = osm_main_roads(tc_data_osm)
junctions = osm_get_junctions(x)
junction_polygons_15 = cluster_junction(junctions)
junction_polygons_30 = cluster_junction(junctions, dist = 30)
plot(x$geometry, col = "grey")
plot(junction_polygons_30, add = TRUE)
plot(junction_polygons_15, add = TRUE)
}
\seealso{
Other OSM: 
\code{\link{line_segment_sf}()},
\code{\link{nn_line}()},
\code{\link{nn_point}()},
\code{\link{osm_consolidate}()},
\code{\link{osm_get_junctions}()},
\code{\link{osm_main_roads}()}
}
\concept{OSM}
