% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{nn_line}
\alias{nn_line}
\title{Fast search for the nearest lines to a set of points}
\usage{
nn_line(point, lines, k = 50, ncores = 1)
}
\arguments{
\item{point}{a SF data frame of POINTS with projected CRS}

\item{lines}{a SF data frame of lines with projected CRS}

\item{k}{integer how many lines to search}

\item{ncores}{integer how many cores to use in parallel processing, default =
1}
}
\value{
Returns an list of ids and distances
}
\description{
Fast search for the nearest lines to a set of points
}
\details{
The nearest line to a point is a non-trivial calculation which in
theory requires checking the distance to every possible line. This function
takes a shortcut by measuring distances to the centroids of lines, and then
cross checking only nearby lines. The number of lines checked is defined by
\code{k}. This process cannot guarantee 100\% accuracy by is many orders of
magnitude faster. To increase accuracy increase \code{k}.
}
\examples{
\dontrun{
crash_road_nn = nn_line(crash_road, osm, ncores = 5)
}
}
\seealso{
Other OSM: 
\code{\link{cluster_junction}()},
\code{\link{line_segment_sf}()},
\code{\link{nn_point}()},
\code{\link{osm_consolidate}()},
\code{\link{osm_get_junctions}()},
\code{\link{osm_main_roads}()}
}
\concept{OSM}
