% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{nn_point}
\alias{nn_point}
\title{Fast search for the nearest point to another set of points}
\usage{
nn_point(x, y, clusters = NULL)
}
\arguments{
\item{x}{a SF data frame of POINTS}

\item{y}{a SF data frame of POINTS}

\item{clusters}{a list of integers, default NULL}
}
\value{
Returns an list of ids and distances
}
\description{
Fast search for the nearest point to another set of points
}
\details{
If \code{clusters} is null will find the nearest y point for each x point
and return a list of two vectors indexes of y length(x) and a distances
length(x). If \code{clusters} is a list e.g. from \code{cluster_junction} then the
indexes are replaced with the matching indexes of \code{clusters}
}
\examples{
\dontrun{
junctions = osm_get_junctions(osm)
junction_clusters = cluster_junction(junctions)

# Find the nearest junction
near_junction = nn_point(crash_junction, junctions)

# Find the nearest junction cluster
near_cluster = nn_point(crash_junction, junctions, clusters = junction_clusters$junction_ids)
}

}
\seealso{
Other OSM: 
\code{\link{cluster_junction}()},
\code{\link{line_segment_sf}()},
\code{\link{nn_line}()},
\code{\link{osm_consolidate}()},
\code{\link{osm_get_junctions}()},
\code{\link{osm_main_roads}()}
}
\concept{OSM}
