% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{osm_consolidate}
\alias{osm_consolidate}
\title{Consolidate roads into simplified network}
\usage{
osm_consolidate(x, segment = 500)
}
\arguments{
\item{x}{a SF data frame of OSM linestrings with projected CRS}

\item{segment}{numeric, max length of segments in metres}
}
\value{
Returns an data frame of LINESTRINGS
}
\description{
Consolidate roads into simplified network
}
\details{
This function simplifies a OSM road network by: 1) grouping small
road segments into single names roads 2) splitting long roads into sections
are defined by \code{segment} 3) Casting MULTILINESTRINGs into LINESTRING

Note: to avoid splitting short roads, roads are only split once they are 2x
\code{segment}, but are then splits into lengths as defined by \code{segment}. For
example a 600m road will not be split, but a 1100m road will be split into
approximately 500m segments.
}
\examples{
\dontrun{
osm = osmextract::oe_read("greater-london-latest.osm.pbf")
osm = osm_main_roads(osm)
osm = sf::st_transform(osm, 27700)
osm = osm_consolidate(osm)
}

}
\seealso{
Other OSM: 
\code{\link{cluster_junction}()},
\code{\link{line_segment_sf}()},
\code{\link{nn_line}()},
\code{\link{nn_point}()},
\code{\link{osm_get_junctions}()},
\code{\link{osm_main_roads}()}
}
\concept{OSM}
