% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{osm_main_roads}
\alias{osm_main_roads}
\title{Select the main roads from OSM}
\usage{
osm_main_roads(
  x,
  highway_values = c("primary", "primary_link", "secondary", "secondary_link",
    "tertiary", "tertiary_link", "trunk", "trunk_link", "motorway", "motorway_link",
    "unclassified", "residential", "road", "mini_roundabout")
)
}
\arguments{
\item{x}{A data frame of OSM lines}

\item{highway_values}{Which highway values to use to define 'main' roads?
Default includes primary, secondar, trunk, motorway, residential and 'mini_roundabout'
road values.}
}
\value{
Returns an data frame
}
\description{
Select the main roads from OSM
}
\details{
The OpenSteetMap contains a lot of detail, this function subsets the
data to just the main roads used by cars by filtering on the highway tag.
}
\examples{
region_name = "Isle of Wight"
# osm = osmextract::oe_get(region_name) # test for IoW
# region_name = "Greater London" # test for London
osm = tc_data_osm
osm_main = osm_main_roads(osm)
nrow(osm)
nrow(osm_main)
nrow(osm_main) / nrow(osm) # keeps ~10-25\% of lines
plot(osm$geometry, col = "grey")
plot(osm_main$geometry, add = TRUE)
}
\seealso{
Other OSM: 
\code{\link{cluster_junction}()},
\code{\link{line_segment_sf}()},
\code{\link{nn_line}()},
\code{\link{nn_point}()},
\code{\link{osm_consolidate}()},
\code{\link{osm_get_junctions}()}
}
\concept{OSM}
