% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_osm.R
\name{tc_get_osm}
\alias{tc_get_osm}
\title{Get traffic calming interventions from OpenStreetMap}
\usage{
tc_get_osm(bbox = NULL, value = NULL, output = "osm_points")
}
\arguments{
\item{bbox}{A bounding box or name passed to \code{osmdata}}

\item{value}{A type of \code{traffic_calming} intervention.
Possible values include \code{yes}, \code{bump} and \code{hump}.
See \href{https://wiki.openstreetmap.org/wiki/Key:traffic_calming}{wiki.openstreetmap.org}
for details.}

\item{output}{The type of \code{osmdata} object to return, \code{osm_points} (default),
\code{osm_lines} etc (see \href{https://cran.r-project.org/package=osmdata}{osmdata} for details)}
}
\value{
An sf object
}
\description{
Get traffic calming interventions from OpenStreetMap
}
\examples{
interventions = tc_get_osm(bbox = "walthamstow village")
}
