% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_join_stats19.R
\name{tc_join_stats19}
\alias{tc_join_stats19}
\title{Basic draft of tc_join_stats19_ac whilst I try to understand issue #24.}
\usage{
tc_join_stats19(
  crashes = NULL,
  casualties = NULL,
  vehicles = NULL,
  level = "accident"
)
}
\arguments{
\item{crashes}{The accidents table of crashes from STATS19 data}

\item{casualties}{The casualty table from STATS19 data}

\item{vehicles}{The vehicles table from STATS19 data}

\item{level}{one of "accidents", "casualities" or "vehicles".
Shorter versions is ok too as a regex of first three letters is used.}
}
\description{
Basic draft of tc_join_stats19_ac whilst I try to understand issue #24.
}
\examples{
\dontrun{
ac = stats19::get_stats19(year = 2019, type = "ac", output_format = "sf")
ca = stats19::get_stats19(year = 2019, type = "ca")
ve = stats19::get_stats19(year = 2019, type = "ve")
tc_join_stats19(ac, ca, veh, level = "veh")
}
}
