% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_join_stats19.R
\name{tc_join_stats19_for_upset}
\alias{tc_join_stats19_for_upset}
\title{Join STATS19 tables to generate UpSet plots with. In the case of vehicle
type UpSet plots, the columns will be types of vehicles and a boolean
to indicate if they were involved in the particular crash. The accident
severity will remain as column too.}
\usage{
tc_join_stats19_for_upset(crashes, casualties, vehicles, pattern_match = NULL)
}
\arguments{
\item{crashes}{The accidents table of crashes from STATS19 data}

\item{casualties}{The casualty table from STATS19 data}

\item{vehicles}{The vehicles table from STATS19 data}

\item{pattern_match}{Character string in the form of
\code{c("car" = "car", "bike" = "bike")} that replaces all parts of the match.}
}
\description{
Join STATS19 tables to generate UpSet plots with. In the case of vehicle
type UpSet plots, the columns will be types of vehicles and a boolean
to indicate if they were involved in the particular crash. The accident
severity will remain as column too.
}
