% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode_casualties}
\alias{tc_recode_casualties}
\title{Recode casualty_type variable from STATS19 data}
\usage{
tc_recode_casualties(x, pattern = NULL, pattern_match = casualties_lookup)
}
\arguments{
\item{x}{Character string to recode}

\item{pattern}{A named character vector with values representing new values.
Has the form \code{c("Car long name" = "Car", "Taxi/Private hire car" = "Taxi")}.
\verb{Car long name} will be converted into \code{Car} in this case.}

\item{pattern_match}{Character string in the form of
\code{c("car" = "car", "bike" = "bike")} that replaces all parts of the match.}
}
\description{
Recode casualty_type variable from STATS19 data
}
\examples{
(x = stats19::casualties_sample$casualty_type)
tc_recode_casualties(x)
\dontrun{
cas = stats19::get_stats19(2018, "casualties")
cas$casualty_type_simple = tc_recode_casualties(cas$casualty_type)
table(cas$casualty_type)
table(cas$casualty_type_simple)
}
}
