% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode_speeds_uk}
\alias{tc_recode_speeds_uk}
\title{Recode maxspeed values in OSM}
\usage{
tc_recode_speeds_uk(
  x,
  allowed_values = c("20 mph", "30 mph", "40 mph", "50 mph", "60 mph", "70 mph"),
  pattern = c(national = "60 mph", `0$` = "0 mph")
)
}
\arguments{
\item{x}{Character string to recode}

\item{allowed_values}{Values that can be returned, other than \code{NA}}

\item{pattern}{A named character vector with values representing new values.
Has the form \code{c("Car long name" = "Car", "Taxi/Private hire car" = "Taxi")}.
\verb{Car long name} will be converted into \code{Car} in this case.}
}
\description{
Recode maxspeed values in OSM
}
\examples{
x = c("30 mph", "10 mph", "60 mph", "25 mph", "30", "national")
tc_recode_speeds_uk(x)
}
