% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode_vehicle_type}
\alias{tc_recode_vehicle_type}
\title{Recode vehicle types}
\usage{
tc_recode_vehicle_type(x)
}
\arguments{
\item{x}{Vehicle types, e.g. from the vehicle_type column in STATS19 data}
}
\description{
This function was designed to simplify vehicle types from STAT19 data
}
\examples{
(x = stats19::vehicles_sample$vehicle_type)
tc_recode_vehicle_type(x)
}
