% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_traffic_calming.R
\name{tc_traffic_calming}
\alias{tc_traffic_calming}
\title{Ge traffic calming measures for an area from OSM}
\usage{
tc_traffic_calming(place = NULL)
}
\arguments{
\item{place}{Geographic name or coordinates as per \code{osmextract}
function \code{oe_get}.}
}
\description{
Use \code{osmextract} to get the roads and then filter on \code{other_tags}
with keyword \code{traffic_calming}
(see \href{https://wiki.openstreetmap.org/wiki/Key:traffic_calming}{OSM Wiki}).
Currently it only takes the \code{place} argument.
}
\examples{
\dontrun{
tc_traffic_calming("Isle of Wight")
}
}
