% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{tc_upset}
\alias{tc_upset}
\title{Create 'who-hit-who' visualisations}
\usage{
tc_upset(crashes_joined)
}
\arguments{
\item{crashes_joined}{crashes dataset with vehicles and casuatlties variables}
}
\description{
Create 'who-hit-who' visualisations
}
