% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cleaning_functions.R
\name{osm_get_junctions}
\alias{osm_get_junctions}
\title{Extract junction points from OSM road linestrings}
\usage{
osm_get_junctions(x)
}
\arguments{
\item{x}{a SF data frame of OSM linestrings}
}
\value{
Returns an SF data frame of POINTS
}
\description{
Extract junction points from OSM road linestrings
}
\details{
This function finds all junction points in a road network, i.e.
where two roads meet. It excludes road crossings e.g. bridges.
}
\examples{
library(sf)
x = osm_main_roads(tc_data_osm)
junctions = osm_get_junctions(x)
plot(x$geometry, col = "grey")
plot(junctions, add = TRUE)
}
\seealso{
Other OSM: 
\code{\link{cluster_junction}()},
\code{\link{line_segment_sf}()},
\code{\link{nn_line}()},
\code{\link{nn_point}()},
\code{\link{osm_consolidate}()},
\code{\link{osm_main_roads}()}
}
\concept{OSM}
