% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_cid.R
\name{tc_get_cid}
\alias{tc_get_cid}
\title{Get traffic calming data from the Cycling Infrastructure Database (CID)}
\usage{
tc_get_cid(type = "traffic_calming", tc_type = NULL, borough = NULL)
}
\arguments{
\item{type}{What type of CID data?
Options include "advanced_stop_line", "crossing", "cycle_lane_track" and}

\item{tc_type}{What type of traffic calming intervention to get?
Options are \code{NULL} (the default, meaning all traffic calming interventions)
or "TRF_RAISED", "TRF_ENTRY", "TRF_CUSHI", "TRF_HUMP", "TRF_SINUSO",}

\item{borough}{Which borough of London is of interest?
\code{NULL} (all boroughs) is the default.}
}
\value{
An sf object representing traffic calming interventions
}
\description{
This function gets data from the CID.
See  for more info.
}
\examples{
# cid_traffic_calming = tc_get_cid()
# summary(cid_traffic_calming)
#   FEATURE_ID            SVDATE           TRF_RAISED      TRF_ENTRY       TRF_CUSHI
# Length:58565       Min.   :2017-01-06   Mode :logical   Mode :logical   Mode :logical
# Class :character   1st Qu.:2017-08-13   FALSE:55793     FALSE:50984     FALSE:45939
# Mode  :character   Median :2017-10-19   TRUE :2772      TRUE :7581      TRUE :12626
# Mean   :2017-10-14
# 3rd Qu.:2017-12-15
# Max.   :2018-08-21
# TRF_HUMP       TRF_SINUSO      TRF_BARIER      TRF_NAROW        TRF_CALM
# Mode :logical   Mode :logical   Mode :logical   Mode :logical   Mode :logical
# FALSE:25294     FALSE:51845     FALSE:57630     FALSE:57903     FALSE:57844
# TRUE :33271     TRUE :6720      TRUE :935       TRUE :662       TRUE :721
#
#
#
# BOROUGH           PHOTO1_URL         PHOTO2_URL                 geometry
# Length:58565       Length:58565       Length:58565       POINT        :58565
# Class :character   Class :character   Class :character   epsg:4326    :    0
# Mode  :character   Mode  :character   Mode  :character   +proj=long...:    0
# cid_barriers = tc_get_cid(tc_type = "TRF_BARIER")
# summary(cid_barriers)
}
