% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_get_osm.R
\name{tc_get_osm}
\alias{tc_get_osm}
\title{Get traffic calming interventions from OpenStreetMap}
\usage{
tc_get_osm(bbox = NULL, value = NULL, output = "osm_points")
}
\arguments{
\item{bbox}{A bounding box or name passed to `osmdata`}

\item{value}{A type of `traffic_calming` intervention.
Possible values include `yes`, `bump` and `hump`.
See [wiki.openstreetmap.org](https://wiki.openstreetmap.org/wiki/Key:traffic_calming)
for details.}

\item{output}{The type of `osmdata` object to return, `osm_points` (default),
`osm_lines` etc (see [osmdata](https://cran.r-project.org/package=osmdata) for details)}
}
\value{
An sf object
}
\description{
Get traffic calming interventions from OpenStreetMap
}
\examples{
interventions = tc_get_osm(bbox = "walthamstow village")
}
