% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_join_stats19.R
\name{tc_join_stats19}
\alias{tc_join_stats19}
\title{Join stats19 tables}
\usage{
tc_join_stats19(crashes, casualties, vehicles, pattern_match = NULL)
}
\arguments{
\item{crashes}{The accidents table of crashes from STATS19 data}

\item{casualties}{The casualty table from STATS19 data}

\item{vehicles}{The vehicles table from STATS19 data}

\item{pattern_match}{Character string in the form of
\code{c("car" = "car", "bike" = "bike")} that replaces all parts of the match.}
}
\description{
Join stats19 tables
}
