% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode}
\alias{tc_recode}
\title{Recode character strings based on regular expressions}
\usage{
tc_recode(x, pattern = NULL, pattern_match = NULL)
}
\arguments{
\item{x}{Character string to recode}

\item{pattern}{A named character vector with values representing new values.
Has the form \code{c("Car long name" = "Car", "Taxi/Private hire car" = "Taxi")}.
\verb{Car long name} will be converted into \code{Car} in this case.}

\item{pattern_match}{Character string in the form of
\code{c("car" = "car", "bike" = "bike")} that replaces all parts of the match.}
}
\description{
This function takes a character string and recodes it.
Default recoding functions are available in functions such as
\code{tc_recode_vehicle_type()} and \code{tc_recode_casualties}.
}
\details{
It is based on the function \code{\link[stringr:str_replace]{stringr::str_replace_all()}}.
}
\examples{
x = c("car long name", "bus", "a bike long", "a bike")
tc_recode(x, pattern = c("car.+" = "car"))
tc_recode(x, pattern = c(".+bike.+" = "bike"))
tc_recode(x, pattern = c("car.+" = "car", ".+bike.+" = "bike"))
tc_recode(x, pattern_match = c("car" = "car", "bike" = "bike"))
}
