% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode_casualties.R
\name{tc_recode_casualties}
\alias{tc_recode_casualties}
\title{Recode casualty types}
\usage{
tc_recode_casualties(x)
}
\arguments{
\item{x}{The type of vehicle the casualty occupied, from the casualty_type column in STATS19 data}
}
\description{
This function was designed to simplify casualty classification from STAT19 data
}
\examples{
(x = stats19::casualties_sample$casualty_type)
(x = stats19::casualties_sample$casualty_type)
tc_recode_casualties(x)
\dontrun{
v = stats19::get_stats19(2018, "casualties")
v$casualty_type_simple = tc_recode_casualties(v$casualty_type)
table(v$casualty_type)
table(v$casualty_type_simple)
}
}
