% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode_vehicle_type}
\alias{tc_recode_vehicle_type}
\title{Recode vehicle types}
\usage{
tc_recode_vehicle_type(
  x,
  pattern = c(`Taxi*.+` = "Taxi", `Van*.+` = "Van", `Pedal cycle` = "Bicycle",
    `(M|m)otorcycle*.+` = "Motorcycle",
    `Data*.+|Elec*.+|Other*.+|Agri*.+|Ridden*.+|Mobility*.+|Tram*.+` = "Other", `Bus*.+`
    = "Bus", `Minibus*.+` = "Minibus", `Goods*.+` = "HGV")
)
}
\arguments{
\item{x}{Vehicle types, e.g. from the vehicle_type column in STATS19 data}

\item{pattern}{A named character vector with values representing new vehicle type names.
Has the form `c("Car" = "Car", "Taxi/Private hire car" = "Taxi")`. See documentation
for details.}
}
\description{
This function was designed to simplify vehicle types from STAT19 data
}
\examples{
(x = stats19::vehicles_sample$vehicle_type)
tc_recode_vehicle_type(x)
\dontrun{
v = stats19::get_stats19(2018, "vehicles")
v$vehicle_type_simple = tc_recode_vehicle_type(v$vehicle_type)
table(v$vehicle_type)
table(v$vehicle_type_simple)
}
}
