% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_recode.R
\name{tc_recode_vehicle_type}
\alias{tc_recode_vehicle_type}
\title{Recode vehicle type}
\usage{
tc_recode_vehicle_type(
  x,
  pattern = c(`Taxi*.+` = "Taxi", `Van*.+` = "Van", `Pedal cycle` = "Bicycle",
    `(M|m)otorcycle*.+|Elec*.+` = "Motorcycle",
    `Data*.+|Other*.+|Agri*.+|Ridden*.+|Mobility*.+|Tram*.+` = "Other", `Bus*.+` = "Bus",
    `Minibus*.+` = "Other", `Goods*.+` = "HGV")
)
}
\arguments{
\item{x}{Character string to recode}

\item{pattern}{A named character vector with values representing new values.
Has the form \code{c("Car long name" = "Car", "Taxi/Private hire car" = "Taxi")}.
\verb{Car long name} will be converted into \code{Car} in this case.}
}
\description{
Recode vehicle type
}
\examples{
(x = stats19::vehicles_sample$vehicle_type)
tc_recode_vehicle_type(x)
\dontrun{
v = stats19::get_stats19(2018, "vehicles")
v$vehicle_type_simple = tc_recode_vehicle_type(v$vehicle_type)
table(v$vehicle_type)
table(v$vehicle_type_simple)
}
}
\seealso{
tc_recode
}
