% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_traffic_calming.R
\name{tc_traffic_calming}
\alias{tc_traffic_calming}
\title{Ge traffic calming measures for an area from OSM}
\usage{
tc_traffic_calming(place = NULL, convert_to_points = FALSE)
}
\arguments{
\item{place}{Geographic name or coordinates as per \code{osmextract}
function \code{oe_get}.}

\item{convert_to_points}{Should linestring geometries be converted to points?
\code{FALSE} by default.}
}
\description{
Use \code{osmextract} to get the roads and then filter on \code{other_tags}
with keyword \code{traffic_calming}
(see \href{https://wiki.openstreetmap.org/wiki/Key:traffic_calming}{OSM Wiki}).
Currently it only takes the \code{place} argument.
}
\examples{
\dontrun{
tc_interventions = tc_traffic_calming("Isle of Wight")
plot(tc_interventions)
tc_interventions
summary(tc_interventions$geometry) # points and lines in there
tc_points = tc_traffic_calming("Isle of Wight", convert_to_points = TRUE)
summary(tc_points$geometry) # points only
tc_wy = tc_traffic_calming("West Yorkshire")
table(tc_wy$traffic_calming)
}
}
