% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{tc_upset}
\alias{tc_upset}
\title{Create 'who-hit-who' visualisations}
\usage{
tc_upset(
  crash_summary,
  casualty_type = c("Car", "Pedestrian", "Van", "Bicycle", "Motorcycle", "Bus"),
  family = ""
)
}
\arguments{
\item{crash_summary}{crashes dataset with vehicles and casuatlties variables}

\item{casualty_type}{The casualty types to plot}

\item{family}{font family}
}
\description{
Create 'who-hit-who' visualisations
}
\examples{
crash_summary = tc_join_stats19(crashes_wf, casualties_wf, vehicles_wf)
tc_upset(crash_summary)
tc_upset(crash_summary, casualty_type = c("Car", "Pedestrian", "Bicycle"))
}
