% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{tc_upset}
\alias{tc_upset}
\title{Create 'who-hit-who' visualisations}
\usage{
tc_upset(
  crash_summary,
  casualty_type = c("Car", "Taxi", "Pedestrian", "Van", "Bicycle", "Motorcycle",
    "Other", "Bus", "HGV", "Minibus"),
  family = "Avenir Book"
)
}
\arguments{
\item{crash_summary}{crashes dataset with vehicles and casuatlties variables}

\item{family}{font family}
}
\description{
Create 'who-hit-who' visualisations
}
\examples{
\dontrun{
crash_summary_wf = tc_join_stats19(crashes, casualties, vehicles)
tc_upset(crash_summary_wy)
}
}
