% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{tc_upset}
\alias{tc_upset}
\title{Create 'who-hit-who' visualisations}
\usage{
tc_upset(
  crash_summary,
  casualty_type = c("Car", "Pedestrian", "Van", "Bicycle", "Motorcycle", "Bus", "HGV",
    "Taxi"),
  family = ""
)
}
\arguments{
\item{crash_summary}{crashes dataset with vehicles and casuatlties variables}

\item{casualty_type}{The casualty types to plot}

\item{family}{font family}
}
\description{
Create 'who-hit-who' visualisations
}
\examples{
crash_summary = tc_join_stats19(crashes_wf, casualties_wf, vehicles_wf)
names(crash_summary)
summary(crash_summary)
tc_upset(crash_summary)
tc_upset(crash_summary, casualty_type = c("Car", "Pedestrian", "Bicycle"))
# create plot with 'Other' category
table(casualties_wf$casualty_type)
casualties_wf2 = dplyr::mutate(
  casualties_wf,
  casualty_type_simple = dplyr::case_when(
     casualty_type == "Car occupant" ~ "Car",
     casualty_type == "Pedestrian" ~ "Pedestrian",
     casualty_type == "Cyclist" ~ "Cyclist",
     TRUE ~ "Other"
    )
  )
table(vehicles_wf$vehicle_type)
vehicles_wf2 = dplyr::mutate(
  vehicles_wf,
  vehicle_type_simple = dplyr::case_when(
     vehicle_type == "Car" ~ "Car",
     vehicle_type == "Cyclist" ~ "Pedal cycle",
     TRUE ~ "Other"
    )
  )
crash_summary = tc_join_stats19(crashes_wf, casualties_wf2, vehicles_wf2)
tc_upset(crash_summary, casualty_type = c("Car", "Pedestrian", "Bicycle", "Other"))
}
